+++ hutool-http/src/main/java/com/xiaoleilu/hutool/http/HttpRequest.java
@@ -497,13 +497,31 @@ public class HttpRequest extends HttpBase<HttpRequest> {
 		}
 		return this;
 	}
+	
+	/**
+	 * 执行Reuqest请求
+	 * 
+	 * @return this
+	 */
+	public HttpResponse execute() {
+		return this.execute(false);
+	}
+	
+	/**
+	 * 异步请求
+	 * @return 异步对象，使用get方法获取HttpResponse对象
+	 */
+	public HttpResponse executeAsync(){
+		return this.execute(true);
+	}
 
 	/**
 	 * 执行Reuqest请求
 	 * 
-	 * @return HttpResponse
+	 * @param isAsync 是否异步
+	 * @return this
 	 */
-	public HttpResponse execute() {
+	public HttpResponse execute(boolean isAsync) {
 		//初始化URL
 		urlWithParamIfGet();
 		// 初始化 connection
@@ -517,17 +535,17 @@ public class HttpRequest extends HttpBase<HttpRequest> {
 
 		// 获取响应
 		if(null == httpResponse){
-			httpResponse = HttpResponse.readResponse(this.httpConnection, this.charset);
+			httpResponse = new HttpResponse(this.httpConnection, this.charset, isAsync);
 		}
-
-		this.httpConnection.disconnect();
 		return httpResponse;
 	}
 	
 	/**
 	 * 异步请求
 	 * @return 异步对象，使用get方法获取HttpResponse对象
+	 * @deprecated 请使用{@link #executeAsync()}
 	 */
+	@Deprecated
 	public Future<HttpResponse> asyncExecute(){
 		return ThreadUtil.execAsync(new Callable<HttpResponse>(){
 			@Override