+++ hutool-http/src/main/java/com/xiaoleilu/hutool/http/HttpResponse.java
@@ -6,7 +6,6 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.Charset;
-import java.nio.charset.UnsupportedCharsetException;
 import java.util.List;
 import java.util.Map.Entry;
 import java.util.zip.GZIPInputStream;
@@ -17,13 +16,20 @@ import com.xiaoleilu.hutool.io.IoUtil;
 import com.xiaoleilu.hutool.util.StrUtil;
 
 /**
- * Http响应类
+ * Http响应类<br>
+ * 非线程安全对象
  * 
  * @author Looly
  *
  */
 public class HttpResponse extends HttpBase<HttpResponse> {
 	
+	/** 持有连接对象 */
+	private HttpConnection httpConnection;
+	/** Http请求原始流 */
+	private InputStream in;
+	/** 是否异步，异步下只持有流，否则将在初始化时直接读取body内容 */
+	private volatile boolean isAsync;
 	/** 读取服务器返回的流保存至内存 */
 	private FastByteArrayOutputStream out;
 	/** 响应状态码 */
@@ -31,17 +37,17 @@ public class HttpResponse extends HttpBase<HttpResponse> {
 
 	/**
 	 * 构造
-	 */
-	protected HttpResponse() {
-	}
-	
-	/**
-	 * 构造
+	 * 
+	 * @param httpConnection {@link HttpConnection}
 	 * @param charset 编码
+	 * @param isAsync 是否异步
 	 * @since 3.0.9
 	 */
-	protected HttpResponse(Charset charset) {
+	protected HttpResponse(HttpConnection httpConnection, Charset charset, boolean isAsync) {
+		this.httpConnection = httpConnection;
 		this.charset = charset;
+		this.isAsync = isAsync;
+		init();
 	}
 	
 	/**
@@ -52,7 +58,19 @@ public class HttpResponse extends HttpBase<HttpResponse> {
 	public int getStatus() {
 		return status;
 	}
-
+	
+	/**
+	 * 同步<br>
+	 * 如果为异步状态，则暂时不读取服务器中响应的内容，而是持有Http链接的{@link InputStream}。<br>
+	 * 当调用此方法时，异步状态转为同步状态，此时从Http链接流中读取body内容并暂存在内容中。如果已经是同步状态，则不进行任何操作。
+	 * 
+	 * @return this
+	 * @throws HttpException IO异常
+	 */
+	public HttpResponse sync() throws HttpException{
+		return this.isAsync ? forceSync() : this;
+	}
+	
 	// ---------------------------------------------------------------- Http Response Header start
 	/**
 	 * 获取内容编码
@@ -66,29 +84,35 @@ public class HttpResponse extends HttpBase<HttpResponse> {
 	 * @return 是否为gzip压缩过的内容
 	 */
 	public boolean isGzip(){
-		String contentEncoding = contentEncoding();
+		final String contentEncoding = contentEncoding();
 		return contentEncoding != null && contentEncoding.equalsIgnoreCase("gzip");
 	}
 	// ---------------------------------------------------------------- Http Response Header end
 	
 	// ---------------------------------------------------------------- Body start
 	/**
-	 * 获得服务区响应流
+	 * 获得服务区响应流<br>
+	 * 异步模式下获取Http原生流，同步模式下获取获取到的在内存中的副本<br>
+	 * 如果想在同步模式下获取流，请先调用{@link #sync()}方法强制同步
+	 * 
 	 * @return 响应流
 	 */
 	public InputStream bodyStream(){
+		if(isAsync) {
+			return this.in;
+		}
 		return new ByteArrayInputStream(this.out.toByteArray());
 	}
 	
 	/**
-	 * 获取响应流字节码
+	 * 获取响应流字节码<br>
+	 * 此方法会转为同步模式
+	 * 
 	 * @return byte[]
 	 */
 	public byte[] bodyBytes() {
-		if(null == this.out){
-			return null;
-		}
-		return this.out.toByteArray();
+		sync();
+		return (null == this.out) ? null : this.out.toByteArray();
 	}
 
 	/**
@@ -98,7 +122,7 @@ public class HttpResponse extends HttpBase<HttpResponse> {
 	 */
 	public String body() throws HttpException{
 		try {
-			return HttpUtil.getString(bodyStream(), this.charset, null == this.charset);
+			return HttpUtil.getString(bodyBytes(), this.charset, null == this.charset);
 		} catch (IOException e) {
 			throw new HttpException(e);
 		}
@@ -108,62 +132,52 @@ public class HttpResponse extends HttpBase<HttpResponse> {
 	@Override
 	public String toString() {
 		StringBuilder sb = StrUtil.builder();
-		sb.append("Request Headers: ").append(StrUtil.CRLF);
+		sb.append("Response Headers: ").append(StrUtil.CRLF);
 		for (Entry<String, List<String>> entry : this.headers.entrySet()) {
 			sb.append("    ").append(entry).append(StrUtil.CRLF);
 		}
 		
-		sb.append("Request Body: ").append(StrUtil.CRLF);
+		sb.append("Response Body: ").append(StrUtil.CRLF);
 		sb.append("    ").append(this.body()).append(StrUtil.CRLF);
 		
 		return sb.toString();
 	}
 	
-	// ---------------------------------------------------------------- Protected method start
+	// ---------------------------------------------------------------- Private method start
 	/**
-	 * 读取响应信息
+	 * 初始化Http响应<br>
+	 * 初始化包括：
+	 * <pre>
+	 * 1、读取Http状态
+	 * 2、读取头信息
+	 * 3、持有Http流，并不关闭流
+	 * </pre>
 	 * 
-	 * @param httpConnection Http连接对象
-	 * @param requestCharset 请求编码
 	 * @return this
+	 * @throws HttpException IO异常
 	 */
-	protected static HttpResponse readResponse(HttpConnection httpConnection, Charset requestCharset) {
-		final HttpResponse httpResponse = new HttpResponse(requestCharset);
-		
-		InputStream in = null;
+	private HttpResponse init() throws HttpException{
 		try {
-			httpResponse.status = httpConnection.responseCode();
-			httpResponse.headers =  httpConnection.headers();
-			final String charset = httpConnection.charset();
-			if(StrUtil.isNotBlank(charset)) {
-				try {
-					httpResponse.charset = Charset.forName(charset);
-				} catch (UnsupportedCharsetException e) {
-					//ignore
-				}
+			this.status = httpConnection.responseCode();
+			this.headers = httpConnection.headers();
+			final Charset charset = httpConnection.getCharset();
+			if(null != charset) {
+				this.charset = charset;
 			}
 			
-			if(httpResponse.status < HttpStatus.HTTP_BAD_REQUEST){
-				in = httpConnection.getInputStream();
-			}else{
-				in = httpConnection.getErrorStream();
-			}
-			httpResponse.readBody(in);
+			this.in = (this.status < HttpStatus.HTTP_BAD_REQUEST) ? httpConnection.getInputStream() : httpConnection.getErrorStream();
 		} catch (IOException e) {
 			if(e instanceof FileNotFoundException){
 				//服务器无返回内容，忽略之
 			}else{
 				throw new HttpException(e.getMessage(), e);
 			}
-		}finally{
-			IoUtil.close(in);
 		}
 		
-		return httpResponse;
+		//同步情况下强制同步
+		return this.isAsync ? this : forceSync();
 	}
-	// ---------------------------------------------------------------- Protected method end
 	
-	// ---------------------------------------------------------------- Private method start
 	/**
 	 * 读取主体
 	 * @param in 输入流
@@ -171,7 +185,7 @@ public class HttpResponse extends HttpBase<HttpResponse> {
 	 * @throws IOException
 	 */
 	private void readBody(InputStream in) throws IOException{
-		if(isGzip()){
+		if(isGzip() && false == (in instanceof GZIPInputStream)){
 			in = new GZIPInputStream(in);
 		}
 		
@@ -182,6 +196,39 @@ public class HttpResponse extends HttpBase<HttpResponse> {
 		} catch (EOFException e) {
 			//忽略读取HTTP流中的EOF错误
 		}
+	}
+	
+	/**
+	 * 强制同步，用于初始化<br>
+	 * 强制同步后变化如下：
+	 * <pre>
+	 * 1、读取body内容到内存
+	 * 2、异步状态设为false（变为同步状态）
+	 * 3、关闭Http流
+	 * 4、断开与服务器连接
+	 * </pre>
+	 * 
+	 * @return this
+	 */
+	private HttpResponse forceSync() {
+		//非同步状态转为同步状态
+		try {
+			this.readBody(this.in);
+		} catch (IOException e) {
+			if(e instanceof FileNotFoundException){
+				//服务器无返回内容，忽略之
+			}else{
+				throw new HttpException(e);
+			}
+		}finally {
+			if(this.isAsync) {
+				this.isAsync = false;
+			}
+			IoUtil.close(this.in);
+			//关闭连接
+			this.httpConnection.disconnect();
+		}
+		return this;
 	}
 	// ---------------------------------------------------------------- Private method end
 }