+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * 增加Copier接口抽象拷贝
 * 增加FileCopier强化文件拷贝
 * ZipUtil增加多个流的方法（issue#IE5ZC@osc）
+* HttpResponse支持body异步，既执行executeAsync()方法后不再直接读取body，而是持有http流对象
 
 ### Bug修复
 * 修复Entity.parseEntity方法中获取表名逻辑的问题（感谢@【北京】游弋苍茫）
@@ -34,6 +35,7 @@
 * 修复Http部分定义编码不能应用于Response的问题（issue#31@osc）
 * 修复FileUtil.equals方法在两个文件都不存在的情况下判定问题
 * 修复Http请求结果多出一个换行的问题（感谢@【北京】仓山有井名为空）
+* 修复StrUtil.cleanBlank方法中length方法调用两次问题（感谢@【天津】〓下页）
 
 ## 3.0.8